Move stats calculation and student list component
