+++ src/web/app/components/student-list/join-state.pipe.ts
