+++ src/web/app/components/student-list/student-list.component.scss
