+++ src/web/app/components/student-list/student-list.component.spec.ts
@@ -4,7 +4,7 @@ import { MatSnackBarModule } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { JoinState } from '../../../types/api-output';
-import { TeammatesCommonModule } from '../../components/teammates-common/teammates-common.module';
+import { TeammatesCommonModule } from '../teammates-common/teammates-common.module';
 import { JoinStatePipe } from './join-state.pipe';
 import { StudentListComponent } from './student-list.component';
