+++ src/web/app/components/student-list/student-list.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
-import { TeammatesCommonModule } from '../../components/teammates-common/teammates-common.module';
+import { TeammatesCommonModule } from '../teammates-common/teammates-common.module';
 import { JoinStatePipe } from './join-state.pipe';
 import { StudentListComponent } from './student-list.component';
