+++ src/web/app/pages-instructor/instructor-course-details-page/instructor-course-details-page.component.spec.ts
@@ -5,7 +5,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { ClipboardModule } from 'ngx-clipboard';
 import { Course, Instructor, InstructorPermissionRole, JoinState } from '../../../types/api-output';
 import { TeammatesCommonModule } from '../../components/teammates-common/teammates-common.module';
-import { StudentListRowModel } from '../student-list/student-list.component';
+import { StudentListRowModel } from '../../components/student-list/student-list.component';
 import { InstructorCourseDetailsPageComponent } from './instructor-course-details-page.component';
 import { InstructorCourseDetailsPageModule } from './instructor-course-details-page.module';
