+++ src/web/app/pages-instructor/instructor-course-details-page/instructor-course-details-page.module.ts
@@ -4,7 +4,7 @@ import { RouterModule } from '@angular/router';
 import { ClipboardModule } from 'ngx-clipboard';
 import { AjaxPreloadModule } from '../../components/ajax-preload/ajax-preload.module';
 import { TeammatesCommonModule } from '../../components/teammates-common/teammates-common.module';
-import { StudentListModule } from '../student-list/student-list.module';
+import { StudentListModule } from '../../components/student-list/student-list.module';
 import { InstructorCourseDetailsPageComponent } from './instructor-course-details-page.component';
 
 /**
