+++ src/web/app/pages-instructor/instructor-search-page/student-result-table/student-result-table.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { StudentListSectionData } from '../../student-list/student-list-section-data';
-import { StudentListRowModel } from '../../student-list/student-list.component';
+import { StudentListSectionData } from '../../../components/student-list/student-list-section-data';
+import { StudentListRowModel } from '../../../components/student-list/student-list.component';
 
 /**
  * Search result for a list of sections containing students of a course
