+++ src/web/app/pages-instructor/instructor-student-list-page/instructor-student-list-page.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { StudentListModule } from '../student-list/student-list.module';
+import { StudentListModule } from '../../components/student-list/student-list.module';
 import { InstructorStudentListPageComponent } from './instructor-student-list-page.component';
 
 /**
