+++ src/web/services/course.service.ts
@@ -2,9 +2,18 @@ import { Injectable } from '@angular/core';
 import { forkJoin, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { ResourceEndpoints } from '../types/api-endpoints';
-import { Course, CourseArchive, Courses,  HasResponses, JoinStatus, MessageOutput } from '../types/api-output';
+import { Course, CourseArchive, Courses, HasResponses, JoinStatus, MessageOutput, Student } from '../types/api-output';
 import { CourseArchiveRequest, CourseCreateRequest, CourseUpdateRequest } from '../types/api-request';
 import { HttpRequestService } from './http-request.service';
+
+/**
+ * The statistics of a course
+ */
+export interface CourseStatistics {
+  numOfSections: number;
+  numOfTeams: number;
+  numOfStudents: number;
+}
 
 /**
  * Handles course related logic provision.
@@ -242,5 +251,22 @@ export class CourseService {
       courseid: courseId,
     };
     return this.httpRequestService.get(ResourceEndpoints.COURSE_SECTIONS, paramsMap);
+  }
+
+  /**
+   * Calculates the statistics for a course from a list of students in the course
+   */
+  calculateCourseStatistics(students: Student[]): CourseStatistics {
+    const teams: Set<string> = new Set();
+    const sections: Set<string> = new Set();
+    students.forEach((student: Student) => {
+      teams.add(student.teamName);
+      sections.add(student.sectionName);
+    });
+    return {
+      numOfSections: sections.size,
+      numOfTeams: teams.size,
+      numOfStudents: students.length,
+    };
   }
 }
